requirements.txt
numpy
pandas
Flask
pytesseract
gunicorn
imutils
#Keras
opencv-python
Pillow
#tensorflow
scikit-image